requirements.txt
altgraph==0.17
dlib==19.18.0
fpdf==1.7.2
imutils==0.5.3
joblib==0.14.1
numpy==1.17.2
opencv-python==4.2.0.32
pandas==0.25.1
Pillow==7.0.0
PyInstaller==3.6
python-dateutil==2.8.0
pytz==2019.3
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
